Move common types to core package. Issue #4
